+++ CHANGELOG.md
@@ -13,8 +13,21 @@
 * Remove hover-over text next to "Effective call number" on the Item record detail view. Refs UIIN-3131.
 * Change import of `exportToCsv` from `stripes-util` to `stripes-components`. Refs UIIN-3025.
 * ECS: Disable opening item details if a user is not affiliated with item's member tenant. Fixes UIIN-3187.
-* Display failure message during `Update Ownership` action when Item contains Local reference data. Fixes UIIN-3195.
 * Correctly depend on `inflected`. Refs UIIN-3203.
+* Decrease the amount of rerenders in `ConsortialHoldings` component. Fixes UIIN-3196.
+* Sort holdings by location name and call number. Refs UIIN-3207.
+* Remove hover-over text next to "Shelving order" on the Item record detail view. Refs UIIN-3210.
+* CI: Switch to centralized/shared workflow from https://github.com/folio-org/.github. Fixes UIIN-3218.
+
+## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
+[Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)
+
+* Detail view of created Instance record is not loaded after saving. Fixes UIIN-3194.
+
+## [12.0.11](https://github.com/folio-org/ui-inventory/tree/v12.0.11) (2025-01-24)
+[Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.10...v12.0.11)
+
+* Display failure message during `Update Ownership` action when Item contains Local reference data. Fixes UIIN-3195.
 
 ## [12.0.10](https://github.com/folio-org/ui-inventory/tree/v12.0.10) (2025-01-20)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.9...v12.0.10)